requirements.txt
eth_account==0.10.0
pycryptodome==3.11.0
web3==6.11.4
